+++ src/config.ts
@@ -14,16 +14,16 @@ import * as utils from './utils';
 export class Config {
 
     // Config settings used on Leo's side
-    public checkForChangeExternalFiles: string = Constants.CONFIG_DEFAULTS.CHECK_FOR_CHANGE_EXTERNAL_FILES;
-    public defaultReloadIgnore: string = Constants.CONFIG_DEFAULTS.DEFAULT_RELOAD_IGNORE;
+    // public checkForChangeExternalFiles: string = Constants.CONFIG_DEFAULTS.CHECK_FOR_CHANGE_EXTERNAL_FILES;
+    // public defaultReloadIgnore: string = Constants.CONFIG_DEFAULTS.DEFAULT_RELOAD_IGNORE;
     public leoID: string = Constants.CONFIG_DEFAULTS.LEO_ID;
-    
+
     //@+others
     //@+node:felix.20260322214546.1: *3* getConfig
     public getConfig(): ConfigMembers {
         return {
-            checkForChangeExternalFiles: this.checkForChangeExternalFiles,
-            defaultReloadIgnore: this.defaultReloadIgnore,
+            // checkForChangeExternalFiles: this.checkForChangeExternalFiles,
+            // defaultReloadIgnore: this.defaultReloadIgnore,
             leoID: this.leoID
         };
     }
@@ -36,12 +36,12 @@ export class Config {
     public setLeoWebSettings(p_changes: ConfigSetting[]): void {
         p_changes.forEach(i_change => {
             switch (i_change.code) {
-                case 'checkForChangeExternalFiles':
-                    this.checkForChangeExternalFiles = i_change.value;
-                    break;
-                case 'defaultReloadIgnore':
-                    this.defaultReloadIgnore = i_change.value;
-                    break;
+                // case 'checkForChangeExternalFiles':
+                //     this.checkForChangeExternalFiles = i_change.value;
+                //     break;
+                // case 'defaultReloadIgnore':
+                //     this.defaultReloadIgnore = i_change.value;
+                //     break;
                 case 'leoID':
                     this.leoID = i_change.value;
                     break;
@@ -54,8 +54,8 @@ export class Config {
         const savedConfig = utils.safeLocalStorageGet(Constants.LOCAL_STORAGE_KEY);
         if (savedConfig !== null) {
             const parsedConfig = JSON.parse(savedConfig);
-            this.checkForChangeExternalFiles = parsedConfig.checkForChangeExternalFiles || this.checkForChangeExternalFiles;
-            this.defaultReloadIgnore = parsedConfig.defaultReloadIgnore || this.defaultReloadIgnore;
+            // this.checkForChangeExternalFiles = parsedConfig.checkForChangeExternalFiles || this.checkForChangeExternalFiles;
+            // this.defaultReloadIgnore = parsedConfig.defaultReloadIgnore || this.defaultReloadIgnore;
             this.leoID = parsedConfig.leoID || this.leoID;
         }